// src/db/types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

export interface VoterRegistrationDeadlines {
  DeadlineByMail: string | null;
  DeadlineInPerson: string | null;
  DeadlineOnline: string | null;
  Description: string | null;
  ElectionDayRegistration: string | null;
  OnlineRegistrationLink: string | null;
  State: string;
}

export interface DB {
  voter_registration_deadlines: VoterRegistrationDeadlines;
}